requirements.txt
awscli==1.33.36
boto3==1.34.154
botocore==1.34.154
colorama==0.4.6
docutils==0.16
jmespath==1.0.1
pyasn1==0.6.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
